speller-lookup: show commits that added or removed looked up words
This uses the new track-words.py utility from the wordlist-diff repo.
As such it two new submodules are added (one for the code branch with
the util, and one for the diff branch).  The init-submodules.sh script
can be used to initialize all the submodules (including the existing
scowl submodule) without bringing in unnecessary history.

+++ update.sh
@@ -7,22 +7,26 @@ set -e
 echo "*** updating"
 
 cd /opt/wl-web-app
-rm -rf git-new
-cp -a -n git git-new
+if [ -d git-old ]; then
+  rm -rf git-new
+  mv git-old git-new
+fi
+
+rsync -aH --delete --exclude=/scowl/scowl.db --exclude=__pycache__ git/ git-new/
 
 cd git-new
 git fetch --recurse-submodules origin deploy
 git reset --recurse-submodules --hard origin/deploy
-git clean -q -f -x -d
+git clean -q -f -x -d -e /history.db
 sh update-esdb.sh
 
 echo "*** copying into place and restarting web-app(s)"
 
 cd /opt/wl-web-app
-rm -rf git-old
 
 systemctl stop wl-web-app-create wl-web-app-speller-lookup
 
+rm -rf git-old
 mv git git-old
 mv git-new git
